api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Andrew Minkin <minkin.andrew@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
Flask = "^2.1.2"
SQLAlchemy = "^1.4.36"
Flask-Migrate = "^3.1.0"
flask-marshmallow = "^0.14.0"
apispec = {extras = ["yaml"], version = "^5.2.2"}
apispec-webframeworks = "^0.5.2"
Flask-RESTful = "^0.3.9"
marshmallow-sqlalchemy = "^0.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
